django/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["hellwebprice <hellwebprice@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.7"
python-socketio = "^5.8.0"
aioredis = "^2.0.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
djoser = "^2.2.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
redis = "^4.6.0"
websockets = "^11.0.3"
pip = "^23.2.1"
django-admin-inline-paginator = "^0.4.0"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
